ramips: add Sercomm partition map parser
This adds an MTD partition parser for the Sercomm partition table that
is used in some Netgear routers.

This is essentially the same code as proposed in the pull request for
Netgear R6350 support by NOGUCHI Hiroshi <drvlabo@gmail.com>:
https://github.com/openwrt/openwrt/pull/1318

It was originally rejected as it did not seem to work correctly.
However, this was only due the NAND driver transparently shifting pages
to hide bad blocks, which was fixed in commit
527832e54bf3bc4d699a145ae66f34230246f0a9.

Signed-off-by: Jan Hoffmann <jan@3e8.eu>
[x1@disroot.org: correction from checkpatch.pl]
Signed-off-by: Maximilian Weinmann <x1@disroot.org>